Minor changes to pitch detection lib

+++ pitch_detection/scripts/peaks_analyser.h
@@ -1,3 +1,5 @@
+#include <stdbool.h>
+
 //Number of frames recorded per second (typically 44.1kHz for CD).
 #define FRAME_RATE 2560
 //The number of values each frame can take tp describe the waveform.
@@ -84,7 +86,7 @@ frequency_bin* get_peaks(const complex clip[]){
             && _is_maxima(cabs(clip[f-1]), amplitude, cabs(clip[f+1]))
             && _is_above_threshold(amplitude, noise)
         ){
-            peaks[i][0] = (double) f * FRAME_RATE / CLIP_FRAMES;
+            peaks[i][0] = f * FRAME_RATE / CLIP_FRAMES;
             peaks[i][1] = decibels(amplitude * 2 / CLIP_FRAMES);
             i++;
         }
